Updated namespace Created EBD_fitness_def Created EBD_fitness_fx Created EBD_parent_sampler Created EBD_reproduction

+++ R/EBD_pop.R
@@ -82,7 +82,7 @@ setMethod( "pop.set", signature( pop = "EBD_pop",  phenotypes = "numeric", index
 } )
 
 
-setMethod( "pop.set", signature( pop = "EBD_pop", phenotypes = "missing", index = "numeric", genotypes = "matrix" ), function( pop, index, genotypes ){
+setMethod( "pop.set", signature( pop = "EBD_pop", phenotypes = "missing", index = "integer", genotypes = "matrix" ), function( pop, index, genotypes ){
     EBD.insert_genotype( pop, genotypes, index )
 } )
